LayoutTests/imported/w3c/web-platform-tests/html/canvas/offscreen/text/2d.text.drawing.style.fontVariant.settings-expected.txt

2d.text.drawing.style.fontVariant.settings

Testing basic functionalities of fontKerning for canvas


FAIL Testing basic functionalities of fontKerning for canvas assert_equals: ctx.fontVariantCaps === "normal" (got [undefined], expected normal[string]) expected (string) "normal" but got (undefined) undefined
